# src/main/java/com/adverity/wka/codechallenge/presentation/rest/DataEntryApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.23).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.adverity.wka.codechallenge.presentation.rest;

import javax.validation.Valid;
import java.util.List;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import com.adverity.wka.codechallenge.presentation.rest.model.DataEntrySearchCriteriaPO;
import com.adverity.wka.codechallenge.presentation.rest.model.DataEntrySearchItemPO;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-11-08T18:17:47.379Z[GMT]")
public interface DataEntryApi {

    @Operation(summary = "Search for data entries", description = "", tags = { "dataEntry" })
    @ApiResponses(value = {
            @ApiResponse(responseCode = "200", description = "successful operation", content = @Content(array = @ArraySchema(schema = @Schema(implementation = DataEntrySearchItemPO.class)))),
            @ApiResponse(responseCode = "400", description = "Invalid input") })
    @RequestMapping(value = "/data-entry/search",
                    produces = { "application/json" },
                    consumes = { "application/json" },
                    method = RequestMethod.POST)
    ResponseEntity<List<DataEntrySearchItemPO>> search(@Parameter(in = ParameterIn.DEFAULT, description = "Criteria to be used for searching", required = true, schema = @Schema()) @Valid @RequestBody DataEntrySearchCriteriaPO body);

}